Split out validation tests
