+++ tests/integration/cvs_handling/input4MIPs/test_cv_loading_integration.py
@@ -8,14 +8,12 @@ see `test_raw_cv_loading.py`.
 from __future__ import annotations
 
 import os
-import re
 from pathlib import Path
 from unittest.mock import patch
 
 import pytest
 from attrs import evolve
 
-from input4mips_validation.cvs_handling.exceptions import NotURLError
 from input4mips_validation.cvs_handling.input4MIPs.activity_id import (
     ActivityIDEntries,
     ActivityIDEntry,
@@ -138,31 +136,3 @@ def test_load_cvs(input4mips_cv_source, exp_except_source):
         res = load_cvs()
 
     assert res == exp
-
-
-def test_activity_id_is_not_url_error():
-    start = load_cvs(
-        raw_cvs_loader=get_raw_cvs_loader(cv_source=DEFAULT_TEST_INPUT4MIPS_CV_SOURCE)
-    )
-
-    inp = evolve(
-        start,
-        activity_id_entries=ActivityIDEntries(
-            (
-                ActivityIDEntry(
-                    activity_id="CMIP",
-                    values=ActivityIDValues(
-                        long_name="Some string",
-                        url="Obviously not a URL",
-                    ),
-                ),
-            ),
-        ),
-    )
-
-    error_msg = re.escape(
-        "url for activity_id entry 'CMIP' has a value of 'Obviously not a URL'. "
-        "This should be a URL (use `www.tbd.invalid` as a placeholder if you need)."
-    )
-    with pytest.raises(NotURLError, match=error_msg):
-        assert_cvs_are_valid(inp)
